refactor: split test files by resource in attachment, sharedfile, and user packages
