+++ internal/attachment/service_test.go
@@ -165,741 +165,6 @@ func TestSpaceAttachmentService_Upload(t *testing.T) {
 	}
 }
 
-func TestWikiAttachmentService_Attach(t *testing.T) {
-	cases := map[string]struct {
-		wikiID        int
-		attachmentIDs []int
-
-		expectError bool
-		want        []*model.Attachment
-
-		mockPostFn func(ctx context.Context, spath string, form url.Values) (*http.Response, error)
-	}{
-		"success-single": {
-			wikiID:        1234,
-			attachmentIDs: []int{2},
-			want:          newTestAttachmentSingleList(),
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				assert.Equal(t, "wikis/1234/attachments", spath)
-
-				v := form
-				assert.Equal(t, []string{"2"}, v["attachmentId[]"])
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.SingleListJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"success-multiple": {
-			wikiID:        1,
-			attachmentIDs: []int{2, 5},
-			want:          newTestAttachmentList(),
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.ListJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-wikiID-invalid": {
-			wikiID:        0,
-			attachmentIDs: []int{1, 2},
-			expectError:   true,
-			mockPostFn:    mock.NewUnexpectedPostFn(t),
-		},
-
-		"error-attachmentIDs-invalid": {
-			wikiID:        1,
-			attachmentIDs: []int{0, 1, 2},
-			expectError:   true,
-			mockPostFn:    mock.NewUnexpectedPostFn(t),
-		},
-
-		"error-attachmentIDs-empty": {
-			wikiID:        1,
-			attachmentIDs: []int{},
-			expectError:   true,
-			mockPostFn:    mock.NewUnexpectedPostFn(t),
-		},
-
-		"error-client": {
-			wikiID:        1234,
-			attachmentIDs: []int{2},
-			expectError:   true,
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			wikiID:        1234,
-			attachmentIDs: []int{2},
-			expectError:   true,
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Post = tc.mockPostFn
-			s := attachment.NewWikiService(method)
-
-			attachments, err := s.Attach(context.Background(), tc.wikiID, tc.attachmentIDs)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachments)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachments)
-
-			assert.Len(t, attachments, len(tc.want))
-
-			for i, w := range tc.want {
-				assert.Equal(t, w.ID, attachments[i].ID)
-				assert.Equal(t, w.Name, attachments[i].Name)
-				assert.Equal(t, w.Size, attachments[i].Size)
-				assert.ObjectsAreEqualValues(w.Created, attachments[i].Created)
-			}
-		})
-	}
-}
-
-func TestWikiAttachmentService_List(t *testing.T) {
-	cases := map[string]struct {
-		wikiID int
-
-		expectError bool
-		want        []*model.Attachment
-
-		mockGetFn func(ctx context.Context, spath string, query url.Values) (*http.Response, error)
-	}{
-		"success": {
-			wikiID: 1234,
-			want:   newTestAttachmentList(),
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				assert.Equal(t, "wikis/1234/attachments", spath)
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.ListJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-wikiID-zero": {
-			wikiID:      0,
-			expectError: true,
-			mockGetFn:   mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-wikiID-negative": {
-			wikiID:      -1,
-			expectError: true,
-			mockGetFn:   mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-client": {
-			wikiID:      1234,
-			expectError: true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			wikiID:      1234,
-			expectError: true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Get = tc.mockGetFn
-			s := attachment.NewWikiService(method)
-
-			attachments, err := s.List(context.Background(), tc.wikiID)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachments)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachments)
-
-			assert.Len(t, attachments, len(tc.want))
-
-			for i, w := range tc.want {
-				assert.Equal(t, w.ID, attachments[i].ID)
-				assert.Equal(t, w.Name, attachments[i].Name)
-				assert.Equal(t, w.Size, attachments[i].Size)
-				assert.ObjectsAreEqualValues(w.Created, attachments[i].Created)
-			}
-		})
-	}
-}
-
-func TestWikiAttachmentService_Remove(t *testing.T) {
-	cases := map[string]struct {
-		wikiID       int
-		attachmentID int
-
-		expectError bool
-		want        *model.Attachment
-
-		mockDeleteFn func(ctx context.Context, spath string, form url.Values) (*http.Response, error)
-	}{
-		"success": {
-			wikiID:       1234,
-			attachmentID: 8,
-			want:         newTestAttachment(),
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				assert.Equal(t, "wikis/1234/attachments/8", spath)
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.SingleJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-wikiID-zero": {
-			wikiID:       0,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-wikiID-negative": {
-			wikiID:       -1,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-attachmentID-zero": {
-			wikiID:       1,
-			attachmentID: 0,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-attachmentID-negative": {
-			wikiID:       1,
-			attachmentID: -1,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-client": {
-			wikiID:       1234,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			wikiID:       1234,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Delete = tc.mockDeleteFn
-			s := attachment.NewWikiService(method)
-
-			attachment, err := s.Remove(context.Background(), tc.wikiID, tc.attachmentID)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachment)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachment)
-
-			assert.Equal(t, tc.want.ID, attachment.ID)
-			assert.Equal(t, tc.want.Name, attachment.Name)
-			assert.Equal(t, tc.want.Size, attachment.Size)
-			assert.ObjectsAreEqualValues(tc.want.Created, attachment.Created)
-		})
-	}
-}
-
-func TestIssueAttachmentService_List(t *testing.T) {
-	cases := map[string]struct {
-		issueIDOrKey string
-
-		expectError bool
-		want        []*model.Attachment
-
-		mockGetFn func(ctx context.Context, spath string, query url.Values) (*http.Response, error)
-	}{
-		"success": {
-			issueIDOrKey: "1234",
-			want:         newTestAttachmentList(),
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				assert.Equal(t, "issues/1234/attachments", spath)
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.ListJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-invalid-issueIDOrKey": {
-			issueIDOrKey: "0",
-			expectError:  true,
-			mockGetFn:    mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-client": {
-			issueIDOrKey: "1234",
-			expectError:  true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			issueIDOrKey: "1234",
-			expectError:  true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Get = tc.mockGetFn
-			s := attachment.NewIssueService(method)
-
-			attachments, err := s.List(context.Background(), tc.issueIDOrKey)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachments)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachments)
-
-			assert.Len(t, attachments, len(tc.want))
-
-			for i, w := range tc.want {
-				assert.Equal(t, w.ID, attachments[i].ID)
-				assert.Equal(t, w.Name, attachments[i].Name)
-				assert.Equal(t, w.Size, attachments[i].Size)
-				assert.Equal(t, w.Created, attachments[i].Created)
-			}
-		})
-	}
-}
-
-func TestIssueAttachmentService_Remove(t *testing.T) {
-	cases := map[string]struct {
-		issueIDOrKey string
-		attachmentID int
-
-		expectError bool
-		want        *model.Attachment
-
-		mockDeleteFn func(ctx context.Context, spath string, form url.Values) (*http.Response, error)
-	}{
-		"success": {
-			issueIDOrKey: "1234",
-			attachmentID: 8,
-			want:         newTestAttachment(),
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				assert.Equal(t, "issues/1234/attachments/8", spath)
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.SingleJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-empty-issueKey": {
-			issueIDOrKey: "",
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-attachmentID-zero": {
-			issueIDOrKey: "test",
-			attachmentID: 0,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-client": {
-			issueIDOrKey: "1234",
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			issueIDOrKey: "1234",
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Delete = tc.mockDeleteFn
-			s := attachment.NewIssueService(method)
-
-			attachment, err := s.Remove(context.Background(), tc.issueIDOrKey, tc.attachmentID)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachment)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachment)
-
-			assert.Equal(t, tc.want.ID, attachment.ID)
-			assert.Equal(t, tc.want.Name, attachment.Name)
-			assert.Equal(t, tc.want.Size, attachment.Size)
-			assert.Equal(t, tc.want.Created, attachment.Created)
-		})
-	}
-}
-
-func TestPullRequestAttachmentService_List(t *testing.T) {
-	cases := map[string]struct {
-		projectIDOrKey     string
-		repositoryIDOrName string
-		prNumber           int
-
-		expectError bool
-		want        []*model.Attachment
-
-		mockGetFn func(ctx context.Context, spath string, query url.Values) (*http.Response, error)
-	}{
-		"success": {
-			projectIDOrKey:     "TEST",
-			repositoryIDOrName: "test",
-			prNumber:           1234,
-			want:               newTestAttachmentList(),
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				assert.Equal(
-					t,
-					"projects/TEST/git/repositories/test/pullRequests/1234/attachments",
-					spath,
-				)
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.ListJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-invalid-project": {
-			projectIDOrKey:     "0",
-			repositoryIDOrName: "1",
-			prNumber:           1,
-			expectError:        true,
-			mockGetFn:          mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-invalid-repository": {
-			projectIDOrKey:     "1",
-			repositoryIDOrName: "0",
-			prNumber:           1,
-			expectError:        true,
-			mockGetFn:          mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-invalid-prNumber": {
-			projectIDOrKey:     "1",
-			repositoryIDOrName: "1",
-			prNumber:           0,
-			expectError:        true,
-			mockGetFn:          mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-client": {
-			projectIDOrKey:     "1234",
-			repositoryIDOrName: "test",
-			prNumber:           10,
-			expectError:        true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			projectIDOrKey:     "1234",
-			repositoryIDOrName: "test",
-			prNumber:           10,
-			expectError:        true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Get = tc.mockGetFn
-			s := attachment.NewPullRequestService(method)
-
-			attachments, err := s.List(context.Background(),
-				tc.projectIDOrKey,
-				tc.repositoryIDOrName,
-				tc.prNumber,
-			)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachments)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachments)
-
-			assert.Len(t, attachments, len(tc.want))
-
-			for i, w := range tc.want {
-				assert.Equal(t, w.ID, attachments[i].ID)
-				assert.Equal(t, w.Name, attachments[i].Name)
-				assert.Equal(t, w.Size, attachments[i].Size)
-				assert.Equal(t, w.Created, attachments[i].Created)
-			}
-		})
-	}
-}
-
-func TestPullRequestAttachmentService_Remove(t *testing.T) {
-	cases := map[string]struct {
-		projectIDOrKey     string
-		repositoryIDOrName string
-		prNumber           int
-		attachmentID       int
-
-		expectError bool
-		want        *model.Attachment
-
-		mockDeleteFn func(ctx context.Context, spath string, form url.Values) (*http.Response, error)
-	}{
-		"success": {
-			projectIDOrKey:     "TEST",
-			repositoryIDOrName: "test",
-			prNumber:           1234,
-			attachmentID:       8,
-			want:               newTestAttachment(),
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				assert.Equal(
-					t,
-					"projects/TEST/git/repositories/test/pullRequests/1234/attachments/8",
-					spath,
-				)
-
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.Attachment.SingleJSON)),
-					),
-				}, nil
-			},
-		},
-
-		"error-invalid-project": {
-			projectIDOrKey:     "0",
-			repositoryIDOrName: "test",
-			prNumber:           1,
-			attachmentID:       8,
-			expectError:        true,
-			mockDeleteFn:       mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-invalid-repository": {
-			projectIDOrKey:     "1",
-			repositoryIDOrName: "",
-			prNumber:           1,
-			attachmentID:       8,
-			expectError:        true,
-			mockDeleteFn:       mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-invalid-prNumber": {
-			projectIDOrKey:     "1",
-			repositoryIDOrName: "test",
-			prNumber:           0,
-			attachmentID:       8,
-			expectError:        true,
-			mockDeleteFn:       mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-invalid-attachmentID": {
-			projectIDOrKey:     "1",
-			repositoryIDOrName: "test",
-			prNumber:           1,
-			attachmentID:       0,
-			expectError:        true,
-			mockDeleteFn:       mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-client": {
-			projectIDOrKey:     "1234",
-			repositoryIDOrName: "test",
-			prNumber:           10,
-			attachmentID:       8,
-			expectError:        true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			projectIDOrKey:     "1234",
-			repositoryIDOrName: "test",
-			prNumber:           10,
-			attachmentID:       8,
-			expectError:        true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body: io.NopCloser(
-						bytes.NewReader([]byte(fixture.InvalidJSON)),
-					),
-				}, nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Delete = tc.mockDeleteFn
-			s := attachment.NewPullRequestService(method)
-
-			attachment, err := s.Remove(
-				context.Background(),
-				tc.projectIDOrKey,
-				tc.repositoryIDOrName,
-				tc.prNumber,
-				tc.attachmentID,
-			)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachment)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachment)
-
-			assert.Equal(t, tc.want.ID, attachment.ID)
-			assert.Equal(t, tc.want.Name, attachment.Name)
-			assert.Equal(t, tc.want.Size, attachment.Size)
-			assert.Equal(t, tc.want.Created, attachment.Created)
-		})
-	}
-}
-
 func Test_contextPropagation(t *testing.T) {
 	type ctxKey struct{}
 	sentinel := &struct{}{}
